remove LinkCard component and move all of its logic into SortableLink, refactor Home.tsx component and update types, props
