--- client/src/components/LinkCard/index.ts
@@ -1,3 +1,0 @@
-import LinkCard from './LinkCard';
-
-export default LinkCard;
