+++ client/src/components/SortableLink/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SortableLink';
